% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cx_read_xpt}
\alias{cx_read_xpt}
\title{Read a SAS XPT file (v5 or v8) into a data frame}
\usage{
cx_read_xpt(path)
}
\arguments{
\item{path}{Path to the .xpt file.}
}
\value{
A data.frame with attributes: \code{var_labels}, \code{var_formats}, \code{var_lengths}, \code{dataset_name}, \code{dataset_label}.
}
\description{
Read a SAS XPT file (v5 or v8) into a data frame
}
\keyword{internal}
