% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_var_length}
\alias{get_var_length}
\title{Get variable byte lengths from a data frame}
\usage{
get_var_length(df)
}
\arguments{
\item{df}{A \code{data.frame}.}
}
\value{
Named integer vector of variable byte lengths.
}
\description{
Get variable byte lengths from a data frame
}
